Move item statuses to a separate file
